refactor: improve fullscreen_media, minor changes.
