+++ packages/stream_chat_flutter/lib/src/fullscreen_media/full_screen_media.dart
@@ -2,14 +2,11 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:chewie/chewie.dart';
-import 'package:contextmenu/contextmenu.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:photo_view/photo_view.dart';
-import 'package:stream_chat_flutter/platform_widget_builder/platform_widget_builder.dart';
 import 'package:stream_chat_flutter/src/attachment/thumbnail/media_attachment_thumbnail.dart';
-import 'package:stream_chat_flutter/src/context_menu_items/download_menu_item.dart';
 import 'package:stream_chat_flutter/src/fullscreen_media/full_screen_media_widget.dart';
+import 'package:stream_chat_flutter/src/fullscreen_media/gallery_navigation_item.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:video_player/video_player.dart';
 
@@ -291,22 +288,16 @@ class _FullScreenMediaState extends State<StreamFullScreenMedia> {
                           ? StreamChannelHeaderTheme.of(context).color
                           : Colors.black,
                       duration: kThemeAnimationDuration,
-                      child: ContextMenuArea(
-                        verticalPadding: 0,
-                        builder: (_) => [
-                          DownloadMenuItem(attachment: attachment),
-                        ],
-                        child: PhotoView.customChild(
-                          maxScale: PhotoViewComputedScale.covered,
-                          minScale: PhotoViewComputedScale.contained,
-                          backgroundDecoration: const BoxDecoration(
-                            color: Colors.transparent,
-                          ),
-                          child: StreamMediaAttachmentThumbnail(
-                            media: attachment,
-                            width: double.infinity,
-                            height: double.infinity,
-                          ),
+                      child: PhotoView.customChild(
+                        maxScale: PhotoViewComputedScale.covered,
+                        minScale: PhotoViewComputedScale.contained,
+                        backgroundDecoration: const BoxDecoration(
+                          color: Colors.transparent,
+                        ),
+                        child: StreamMediaAttachmentThumbnail(
+                          media: attachment,
+                          width: double.infinity,
+                          height: double.infinity,
                         ),
                       ),
                     ),
@@ -318,92 +309,16 @@ class _FullScreenMediaState extends State<StreamFullScreenMedia> {
                       child: CircularProgressIndicator.adaptive(),
                     );
                   }
+
                   return InkWell(
                     onTap: switchDisplayingDetail,
-                    child: Padding(
-                      padding: const EdgeInsets.symmetric(vertical: 50),
-                      child: ContextMenuArea(
-                        verticalPadding: 0,
-                        builder: (_) => [
-                          DownloadMenuItem(attachment: attachment),
-                        ],
-                        child: Chewie(
-                          controller: controller.chewieController!,
-                        ),
-                      ),
+                    child: Chewie(
+                      controller: controller.chewieController!,
                     ),
                   );
                 }
                 return const SizedBox();
               },
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// A widget for desktop and web users to be able to navigate left and right
-/// through a gallery of images.
-class GalleryNavigationItem extends StatelessWidget {
-  /// Builds a [GalleryNavigationItem].
-  const GalleryNavigationItem({
-    super.key,
-    required this.icon,
-    this.iconSize = 48,
-    required this.onPressed,
-    required this.opacityAnimation,
-    this.left,
-    this.right,
-  });
-
-  /// The icon to display.
-  final Widget icon;
-
-  /// The size of the icon.
-  ///
-  /// Defaults to 48.
-  final double iconSize;
-
-  /// The callback to perform when the button is clicked.
-  final VoidCallback onPressed;
-
-  /// The animation for showing & hiding this widget.
-  final ValueListenable<bool> opacityAnimation;
-
-  /// The left-hand placement of the button.
-  final double? left;
-
-  /// The right-hand placement of the button.
-  final double? right;
-
-  @override
-  Widget build(BuildContext context) {
-    return PlatformWidgetBuilder(
-      desktop: (_, child) => child,
-      web: (_, child) => child,
-      child: Positioned(
-        left: left,
-        right: right,
-        top: MediaQuery.of(context).size.height / 2,
-        child: ValueListenableBuilder<bool>(
-          valueListenable: opacityAnimation,
-          builder: (context, shouldShow, child) {
-            return AnimatedOpacity(
-              opacity: shouldShow ? 1 : 0,
-              duration: kThemeAnimationDuration,
-              child: child,
-            );
-          },
-          child: Material(
-            color: Colors.transparent,
-            type: MaterialType.circle,
-            clipBehavior: Clip.antiAlias,
-            child: IconButton(
-              icon: icon,
-              iconSize: iconSize,
-              onPressed: onPressed,
             ),
           ),
         ),
